task_manager_bot/requirements.txt
discord.py
pytest
pytest-asyncio
requests
aiohttp
cryptography